updated to vs2012, with some minor tweaks to reduce build warnings

+++ DistThresholdCHOP.cpp
@@ -223,7 +223,7 @@ DistThresholdCHOP::getInfoCHOPChan(int index,
 	if (index == 0)
 	{
 		chan->name = "executeCount";
-		chan->value = myExecuteCount;
+		chan->value = (float)myExecuteCount;
 	}
 }
